♻️ 그룹 생성 폼 컴포넌트 분리
