+++ app/frontend/src/pages/Groups/Create/types.ts
@@ -1,0 +1,6 @@
+export type GroupCreate = {
+  name: string;
+  type: 'public' | 'private';
+  joinType: GroupJoin[];
+};
+export type GroupJoin = 'approve' | 'code';
